+++ gstd/src/common/handlers.rs
