+++ gstd/src/common/primitives.rs
