+++ gstd/src/macros/mod.rs
@@ -20,3 +20,26 @@ mod bail;
 mod debug;
 mod export;
 mod metadata;
+
+pub mod util {
+    use crate::prelude::{Box, String, Vec};
+    use codec::Encode;
+    use scale_info::{PortableRegistry, Registry};
+
+    pub use scale_info::MetaType;
+
+    pub fn to_hex_registry(meta_types: Vec<MetaType>) -> String {
+        let mut registry = Registry::new();
+        registry.register_types(meta_types);
+    
+        let registry: PortableRegistry = registry.into();
+        hex::encode(registry.encode())
+    }
+    
+    pub fn to_wasm_ptr<T: AsRef<[u8]>>(bytes: T) -> *mut [i32; 2] {
+        Box::into_raw(Box::new([
+            bytes.as_ref().as_ptr() as _,
+            bytes.as_ref().len() as _,
+        ]))
+    }
+}
