+++ gstd/src/sync/access.rs
@@ -17,27 +17,27 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 use crate::prelude::VecDeque;
-use crate::MessageId;
+use crate::common::MessageId;
 use core::cell::UnsafeCell;
 
 // Option<VecDeque> to make new `const fn`
-pub(crate) struct AccessQueue(UnsafeCell<Option<VecDeque<MessageId>>>);
+pub struct AccessQueue(UnsafeCell<Option<VecDeque<MessageId>>>);
 
 impl AccessQueue {
-    pub(crate) fn enqueue(&self, message_id: MessageId) {
+    pub fn enqueue(&self, message_id: MessageId) {
         let inner = unsafe { &mut *self.0.get() };
 
         let vec_deque = inner.get_or_insert_with(VecDeque::new);
         vec_deque.push_back(message_id);
     }
 
-    pub(crate) fn dequeue(&self) -> Option<MessageId> {
+    pub fn dequeue(&self) -> Option<MessageId> {
         let inner = unsafe { &mut *self.0.get() };
 
         inner.as_mut().and_then(|v| v.pop_front())
     }
 
-    pub(crate) const fn new() -> Self {
+    pub const fn new() -> Self {
         AccessQueue(UnsafeCell::new(None))
     }
 }
